# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/requests": {
      "get": {
        "description": "Возвращает список всех заявок на подключение терминалов к магазинам. Доступно для администраторов(ROOT и MANAGER)",
        "operationId": "RequestsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список заявок",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": "b0816217-c88f-402f-9c26-de995d57ad86",
                      "terminalId": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                      "shopLogin": "testshop1",
                      "status": "PENDING",
                      "comment": "Заявка проверена, все документы в порядке",
                      "createdAt": "2025-10-19T22:45:37.148Z",
                      "updatedAt": "2025-10-19T22:45:47.884Z",
                      "terminal": {
                        "id": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                        "macAddress": "FF:EE:DD:CC:BB:AA",
                        "status": "ACTIVE"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить список заявок",
        "tags": [
          "Заявки"
        ]
      },
      "post": {
        "description": "Создание заявки на подключение терминала к магазину. Заявка создается в статусе PENDING.",
        "operationId": "RequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Заявка создана",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "b0816217-c88f-402f-9c26-de995d57ad86",
                    "terminalId": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                    "shopLogin": "testshop1",
                    "status": "PENDING",
                    "comment": null,
                    "createdAt": "2025-10-19T22:45:37.148Z",
                    "updatedAt": "2025-10-19T22:45:37.148Z",
                    "terminal": {
                      "id": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                      "macAddress": "FF:EE:DD:CC:BB:AA",
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные или заявка уже существует",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request already exists for this terminal and shop",
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Создать новую заявку",
        "tags": [
          "Заявки"
        ]
      }
    },
    "/requests/{id}": {
      "get": {
        "description": "Возвращает детальную информацию о конкретной заявке по её ID.",
        "operationId": "RequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Данные заявки",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "b0816217-c88f-402f-9c26-de995d57ad86",
                    "terminalId": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                    "shopLogin": "testshop1",
                    "status": "PENDING",
                    "comment": null,
                    "createdAt": "2025-10-19T22:45:37.148Z",
                    "updatedAt": "2025-10-19T22:45:37.148Z",
                    "terminal": {
                      "id": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                      "macAddress": "FF:EE:DD:CC:BB:AA",
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Получить заявку по ID",
        "tags": [
          "Заявки"
        ]
      },
      "delete": {
        "operationId": "RequestsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка удалена"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Удалить заявку",
        "tags": [
          "Заявки"
        ]
      }
    },
    "/requests/{id}/approve": {
      "patch": {
        "description": "Одобрение заявки на подключение. Терминал автоматически привязывается к магазину (атомарная операция в транзакции). Статус заявки меняется на APPROVED.",
        "operationId": "RequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка одобрена, терминал подключен к магазину",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request approved and terminal connected to shop",
                    "status": "APPROVED"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Заявка уже обработана",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request has already been processed",
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Одобрить заявку",
        "tags": [
          "Заявки"
        ]
      }
    },
    "/requests/{id}/reject": {
      "patch": {
        "description": "Отклонение заявки на подключение терминала. Статус заявки меняется на REJECTED. Терминал остается не привязанным к магазину.",
        "operationId": "RequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Заявка отклонена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request rejected",
                    "status": "REJECTED"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Заявка уже обработана",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request has already been processed",
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Отклонить заявку",
        "tags": [
          "Заявки"
        ]
      }
    },
    "/requests/{id}/comment": {
      "post": {
        "description": "Добавление или обновление комментария к заявке. Комментарий может содержать причину одобрения/отклонения или дополнительные замечания.",
        "operationId": "RequestsController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Комментарий добавлен",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "b0816217-c88f-402f-9c26-de995d57ad86",
                    "terminalId": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                    "shopLogin": "testshop1",
                    "status": "PENDING",
                    "comment": "Заявка проверена, все документы в порядке",
                    "createdAt": "2025-10-19T22:45:37.148Z",
                    "updatedAt": "2025-10-19T22:45:47.884Z",
                    "terminal": {
                      "id": "0bd492c9-5d78-428d-8ef3-3d5ece1296c6",
                      "macAddress": "FF:EE:DD:CC:BB:AA",
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Request not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Добавить комментарий к заявке",
        "tags": [
          "Заявки"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shop Terminals Management API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен полученный при авторизации",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "terminalId": {
            "type": "string",
            "example": "uuid-terminal-id",
            "description": "ID терминала"
          },
          "shopLogin": {
            "type": "string",
            "example": "shop1",
            "description": "Логин магазина"
          }
        },
        "required": [
          "terminalId",
          "shopLogin"
        ]
      },
      "AddCommentDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "example": "Заявка одобрена после проверки документов",
            "description": "Комментарий к заявке"
          }
        },
        "required": [
          "comment"
        ]
      },
...
    }
  }
}